+++ src/services/ui/src/components/AdditionalInfo/index.tsx
@@ -1,11 +1,11 @@
 export const AdditionalInfo = () => {
   return (
-    <>
-      <p className="mb-2">
-        Thanks, thanks a lot. Is she pretty? After I fell off my toilet, I drew
-        this. Of course, the Enchantment Under The Sea Dance theyre supposed to
-        go to this, thats where they kiss for the first time. Thats Strickland.
-        Jesus, didnt that guy ever have hair?
+    <div className="flex flex-col gap-4">
+      <p>
+        Thanks, thanks a lot. After I fell off my toilet, I drew this. Of
+        course, the Enchantment Under The Sea Dance theyre supposed to go to
+        this, thats where they kiss for the first time. Thats Strickland. Jesus,
+        didnt that guy ever have hair?
       </p>
 
       <p>
@@ -14,9 +14,17 @@ export const AdditionalInfo = () => {
         do you explain that wreck out there? The appropriate question is, werent
         the hell are they. Einstein has just become the worlds first time
         traveler. I sent him into the future. One minute into the future to be
-        exact. And at exactly 1:21 a.m. we should cat h up with him and the time
+        exact. And at exactly 1:21 a.m. we should catch up with him and the time
         machine. Ahh.
       </p>
-    </>
+      <p>
+        Space, the final frontier. These are the voyages of the Starship
+        Enterprise. Its five-year mission: to explore strange new worlds, to
+        seek out new life and new civilizations, to boldly go where no man has
+        gone before. Many say exploration is part of our destiny, but its
+        actually our duty to future generations and their quest to ensure the
+        survival of the human species.
+      </p>
+    </div>
   );
 };
