+++ src/services/ui/src/pages/detail/wrapper.tsx
@@ -1,10 +1,8 @@
 import { SearchData, useGetItem } from "@/api"; // Assuming useGetItem returns RecordData
 import { LoadingSpinner } from "@/components";
-import { ChevronLeftIcon } from "@heroicons/react/24/outline";
-
 import { useQuery } from "@/hooks";
 import React from "react";
-import { useNavigate } from "react-router-dom";
+import { DetailNav } from "./detailNav";
 
 interface DetailWrapperProps {
   children: React.ReactNode;
@@ -16,7 +14,6 @@ interface ChildComponentProps {
 }
 
 export const DetailWrapper: React.FC<DetailWrapperProps> = ({ children }) => {
-  const navigate = useNavigate();
   const query = useQuery();
   const id = query.get("id") as string;
   const { data, isLoading, error } = useGetItem(id);
@@ -30,20 +27,7 @@ export const DetailWrapper: React.FC<DetailWrapperProps> = ({ children }) => {
 
   return (
     <>
-      <div className="bg-sky-100">
-        <div className="max-w-screen-lg m-auto lg:px-8">
-          <div className="flex items-center">
-            <div className="flex align-middle py-4">
-              <button className="text-sky-800" onClick={() => navigate(-1)}>
-                <ChevronLeftIcon className="w-6 h-6" />
-              </button>
-              <h1 className="text-xl font-medium pl-4">
-                CHIP SPA Submission Details - {id}
-              </h1>
-            </div>
-          </div>
-        </div>
-      </div>
+      <DetailNav id={id} />
       <div className="max-w-screen-lg mx-auto py-8 px-4 lg:px-8">
         {React.Children.map(children, (child) => {
           if (React.isValidElement<ChildComponentProps>(child)) {
